pyproject.toml
[tool.poetry]
name = "cryptowallet"
version = "0.0.1"
description = ""
authors = ["SpaceLab"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.2"
fastapi = "^0.85.0"
toml = "^0.10.2"
yarl = "^1.8.1"
email-validator = "^1.3.0"
async-lru = "^1.0.3"
pyhumps = "^3.7.3"
gunicorn = "^20.1.0"
uvicorn = "^0.18.3"
click = "^8.1.3"
httpx = "^0.23.0"
loguru = "^0.6.0"
aioredis = "^2.0.1"
celery = "^5.2.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
